[Sanitizer] Add smoke test for report symbolication
Add a test that checks symbolication for out-of-process and in-process
symbolication for both Apple platforms and Linux.

Note that there already exist tests in the compiler-rt test suite that
check report symbolication.  The purpose of this test is to have simple
smoke tests for Swift, and specifically Swift demangling.

rdar://62753845